refactor(routing): split tests into one file per logical unit
Break datachannel.test.ts and tick-action.test.ts into four focused files:
- datachannel-core-fields.test.ts — required/optional field validation
- datachannel-envoy-port.test.ts — envoyPort field validation
- tick-message-schema.test.ts — TickMessageSchema validation
- action-schema.test.ts — unified ActionSchema and Actions enum
