+++ packages/routing/tests/unit/tick-message-schema.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect } from 'bun:test'
 import { TickMessageSchema } from '../../src/system/actions.js'
-import { ActionSchema } from '../../src/schema.js'
 import { Actions } from '../../src/action-types.js'
 
 describe('TickMessageSchema', () => {
@@ -45,22 +44,5 @@ describe('TickMessageSchema', () => {
       data: { now: Date.now() },
     })
     expect(result.success).toBe(false)
-  })
-})
-
-describe('ActionSchema', () => {
-  it('accepts system:tick via unified ActionSchema', () => {
-    const result = ActionSchema.safeParse({
-      action: 'system:tick',
-      data: { now: 1707745200000 },
-    })
-    expect(result.success).toBe(true)
-    if (result.success) {
-      expect(result.data.action).toBe(Actions.Tick)
-    }
-  })
-
-  it('Actions.Tick equals "system:tick"', () => {
-    expect(Actions.Tick).toBe('system:tick')
   })
 })
